Further work on all scripts necessary for manuscript.
git-svn-id: https://svn.seismo.oma.be/svn/seismo/hazard/psha/trunk/Projects/Aysen_fjord@4537 40b490c5-b4d9-47cb-8714-9bef99b524d5

+++ calc_gmpe_misfit.py
@@ -104,7 +104,7 @@ for gmpe_name in ["AllenEtAl2012", "AtkinsonWald2007", "BakunWentworth1997", "Ba
 					lat, lon = float(lat), float(lon)
 					mmi = {"20070421": mmi1, "20070402": mmi2}[event_ID]
 					try:
-						mmi = int(mmi)
+						mmi = float(mmi)
 					except:
 						pass
 					else:
